+++ src/test/java/com/osoco/challenge/yaus/model/ShortenAlgorithmTest.java
@@ -8,17 +8,17 @@ public class ShortenAlgorithmTest {
 
     @Test
     public void getShortenFor12345ShouldBednh() {
-        assertEquals("dnh", ShortenAlgorithm.getShortUrl(12345));
+        assertEquals("dnh", ShortenAlgorithm.getShortUrl(12345L));
     }
 
     @Test
     public void getIdFordnhShouldBe12345() {
-        assertEquals(12345, ShortenAlgorithm.getId("dnh"));
+        assertEquals(12345L, ShortenAlgorithm.getId("dnh"));
     }
 
     @Test
     public void algorithmShouldBeBijective() {
-        int id = 123456789;
+        long id = 123456789L;
 
         assertEquals(id, ShortenAlgorithm.getId(ShortenAlgorithm.getShortUrl(id)));
     }
